+++ common/src/main/java/it/crystalnest/cobweb/api/pack/dynamic/DynamicResourcePack.java
@@ -1,4 +1,4 @@
-package it.crystalnest.cobweb.api.pack;
+package it.crystalnest.cobweb.api.pack.dynamic;
 
 import com.google.common.base.Suppliers;
 import com.google.gson.JsonElement;