+++ common/src/main/java/it/crystalnest/cobweb/api/pack/dynamic/DynamicResourceType.java
@@ -1,4 +1,4 @@
-package it.crystalnest.cobweb.api.pack;
+package it.crystalnest.cobweb.api.pack.dynamic;
 
 import net.minecraft.resources.ResourceLocation;