+++ fabric/src/main/java/it/crystalnest/cobweb/platform/FabricRegistryHelper.java
@@ -1,5 +1,7 @@
 package it.crystalnest.cobweb.platform;
 
+import it.crystalnest.cobweb.api.pack.PackSources;
+import it.crystalnest.cobweb.api.pack.fixed.StaticResourcePack;
 import it.crystalnest.cobweb.api.registry.CobwebEntry;
 import it.crystalnest.cobweb.api.registry.CobwebRegister;
 import it.crystalnest.cobweb.api.registry.Register;
@@ -11,6 +13,7 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.repository.Pack;
+import net.minecraft.server.packs.repository.PackRepository;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 
@@ -35,21 +38,69 @@ public final class FabricRegistryHelper extends RegistryHelper<FabricRegistryHel
   private static final List<Supplier<Pack>> DYNAMIC_TEXTURE_PACKS = new ArrayList<>();
 
   /**
-   * Returns the list of currently registered dynamic {@link Pack data pack}s.
-   *
-   * @return list of registered dynamic {@link Pack data pack}s.
+   * Internal list of {@link StaticResourcePack} of type {@link PackType#SERVER_DATA}.
    */
-  public static List<Pack> dynamicDataPacks() {
-    return DYNAMIC_DATA_PACKS.stream().map(Supplier::get).toList();
+  private static final List<StaticResourcePack> STATIC_DATA_PACKS = new ArrayList<>();
+
+  /**
+   * Internal list of {@link StaticResourcePack} of type {@link PackType#CLIENT_RESOURCES}.
+   */
+  private static final List<StaticResourcePack> STATIC_TEXTURE_PACKS = new ArrayList<>();
+
+  /**
+   * Registers all data packs, dynamic and static, to the given repository.
+   *
+   * @param repository {@link PackRepository}.
+   * @return updated repository.
+   */
+  public static PackRepository registerDataPacks(PackRepository repository) {
+    registerDynamicPacks(DYNAMIC_DATA_PACKS, repository);
+    registerStaticPacks(STATIC_DATA_PACKS, repository);
+    return repository;
   }
 
   /**
-   * Returns the list of currently registered dynamic {@link Pack texture pack}s.
+   * Registers all texture packs, dynamic and static, to the given repository.
    *
-   * @return list of registered dynamic {@link Pack texture pack}s.
+   * @param repository {@link PackRepository}.
+   * @return updated repository.
    */
-  public static List<Pack> dynamicTexturePacks() {
-    return DYNAMIC_TEXTURE_PACKS.stream().map(Supplier::get).toList();
+  public static PackRepository registerTexturePacks(PackRepository repository) {
+    registerDynamicPacks(DYNAMIC_TEXTURE_PACKS, repository);
+    registerStaticPacks(STATIC_TEXTURE_PACKS, repository);
+    return repository;
+  }
+
+  /**
+   * Registers the provided list of dynamic resource packs to the given repository.
+   *
+   * @param packs list of dynamic packs.
+   * @param repository {@link PackRepository} to update.
+   */
+  private static void registerDynamicPacks(List<Supplier<Pack>> packs, PackRepository repository) {
+    registerPacks(packs, Supplier::get, repository);
+  }
+
+  /**
+   * Registers the provided list of static resource packs to the given repository.
+   *
+   * @param packs list of static packs.
+   * @param repository {@link PackRepository} to update.
+   */
+  private static void registerStaticPacks(List<StaticResourcePack> packs, PackRepository repository) {
+    registerPacks(packs, StaticResourcePack::toPack, repository);
+  }
+
+  /**
+   * Registers the provided list of resource packs to the given repository.
+   *
+   * @param packs resource packs.
+   * @param toPack function to map a resource pack into a {@link Pack}.
+   * @param repository {@link PackRepository}.
+   * @param <T> resource pack type.
+   */
+  private static <T> void registerPacks(List<T> packs, Function<T, Pack> toPack, PackRepository repository) {
+    packs.stream().map(toPack).forEach(pack -> ((PackSources) repository).addSource(packConsumer -> packConsumer.accept(pack)));
   }
 
   @Override
@@ -67,27 +118,35 @@ public final class FabricRegistryHelper extends RegistryHelper<FabricRegistryHel
     return of(DeferredRegister.Blocks::new, Registries.BLOCK, namespace);
   }
 
-  /**
-   * Provides a {@link DeferredRegister} for the specified mod and {@link Registry}.
-   *
-   * @param constructor {@link DeferredRegister} subclass constructor.
-   * @param registryKey Minecraft {@link Registry} key.
-   * @param namespace mod ID.
-   * @return {@link DeferredRegister}.
-   * @param <R> register type.
-   * @param <T> {@link DeferredRegister} type.
-   */
-  @SuppressWarnings("unchecked")
-  private <R, T extends DeferredRegister<R>> T of(Function<String, T> constructor, ResourceKey<? extends Registry<R>> registryKey, String namespace) {
-    return (T) registries.computeIfAbsent(namespace, key -> new HashMap<>()).computeIfAbsent(registryKey.location(), key -> constructor.apply(namespace));
-  }
-
   @Override
   public void registerDynamicResourcePack(PackType type, Supplier<Pack> supplier) {
     switch (type) {
       case SERVER_DATA -> DYNAMIC_DATA_PACKS.add(supplier);
       case CLIENT_RESOURCES -> DYNAMIC_TEXTURE_PACKS.add(supplier);
     }
+  }
+
+  @Override
+  public void registerStaticResourcePack(StaticResourcePack pack) {
+    switch (pack.type()) {
+      case SERVER_DATA -> STATIC_DATA_PACKS.add(pack);
+      case CLIENT_RESOURCES -> STATIC_TEXTURE_PACKS.add(pack);
+    }
+  }
+
+  /**
+   * Provides a {@link DeferredRegister} for the specified mod and {@link Registry}.
+   *
+   * @param constructor {@link DeferredRegister} subclass constructor.
+   * @param registryKey Minecraft {@link Registry} key.
+   * @param namespace mod ID.
+   * @param <R> register type.
+   * @param <T> {@link DeferredRegister} type.
+   * @return {@link DeferredRegister}.
+   */
+  @SuppressWarnings("unchecked")
+  private <R, T extends DeferredRegister<R>> T of(Function<String, T> constructor, ResourceKey<? extends Registry<R>> registryKey, String namespace) {
+    return (T) registries.computeIfAbsent(namespace, key -> new HashMap<>()).computeIfAbsent(registryKey.location(), key -> constructor.apply(namespace));
   }
 
   /**
@@ -116,14 +175,14 @@ public final class FabricRegistryHelper extends RegistryHelper<FabricRegistryHel
     }
 
     @Override
-    @SuppressWarnings("unchecked")
-    public <T extends R> CobwebEntry<T> register(String name, Supplier<? extends T> supplier) {
-      return new CobwebEntry<>(Holder.direct((T) this.register.apply(name, supplier.get())));
+    public String namespace() {
+      return namespace;
     }
 
     @Override
-    public String namespace() {
-      return namespace;
+    @SuppressWarnings("unchecked")
+    public <T extends R> CobwebEntry<T> register(String name, Supplier<? extends T> supplier) {
+      return new CobwebEntry<>(Holder.direct((T) this.register.apply(name, supplier.get())));
     }
 
     /**